без маппера

+++ lit/src/main/java/com/litsite/lit/repository/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<MyUser, Long> {
+public interface UserRepository extends JpaRepository<MyUser, Integer> {
     Optional<MyUser> findByLogin(String login);
     Optional<MyUser> findByUsername(String username);
     Boolean existsByUsername(String username);
